@@ -2,6 +2,8 @@ package com.example.post.web;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface PostRepository extends JpaRepository<Post, Long> {
-    //List<Post> findByAccountNumber(String accountNumber);
+    List<Post> findByAccountNumber(String accountNumber);
 }
